Add group in task and server. And more fixes. Show releases.

+++ www/textfill.h
@@ -6,5 +6,5 @@ using __DP_LIB_NAMESPACE__::List;
 
 String fillText(const String & text, const List<String> & args);
 String findText(const String & text);
-char * findResourc(const std::string & file, unsigned long & size);
+char * findResourc(const std::string & file, unsigned long long & size);
 String findFillText(const String & text, const List<String> & args);
